Use 64-bit arithmetic for tensor buffer size calculations
Buffer size calculations of the form dataType.getNumOfBytes() * size were
performed in 32-bit int. For large element counts these products overflow and
wrap to a small or negative value, so BaseNDManager.validateBuffer compared the
supplied buffer against the wrong expected size, and the corresponding
allocateDirect calls allocated buffers smaller than the shape requires.

Compute these byte counts in 64-bit and narrow with Math.toIntExact so a request
that does not fit in an int throws instead of silently wrapping. The same
treatment is applied to the sibling allocation paths in the api module and in
each engine, and to the (int) shape.size() narrowings in NDManager.zeros/ones
and BaseNDManager.truncatedNormal, which could truncate rather than fail.

Behaviour is unchanged for any shape that fits in an int-sized buffer.

Adds regression coverage for the overflowing element counts.

+++ api/src/main/java/ai/djl/ndarray/NDArrayAdapter.java
@@ -169,7 +169,9 @@ public abstract class NDArrayAdapter implements NDArray {
     }
 
     private ByteBuffer toTypeInternal(Number[] numbers, DataType dataType) {
-        int size = dataType.getNumOfBytes() * numbers.length;
+        // 64-bit arithmetic with a fail-closed narrowing: getNumOfBytes() * length can overflow
+        // int for large arrays, silently allocating an undersized direct buffer.
+        int size = Math.toIntExact((long) dataType.getNumOfBytes() * numbers.length);
         ByteBuffer bb = manager.allocateDirect(size);
         for (Number number : numbers) {
             switch (dataType) {
